flask_redis_postgres/app/requirements.txt
Flask
redis
psycopg2-binary
celery